textractlambda/requirements.txt
amazon-textract-response-parser
amazon-textract-prettyprinter